+++ visualization/app/codeCharta/state/effects/nodeContextMenu/flattenButtons/flattenButtons.component.html
@@ -4,10 +4,10 @@
 	(click)="flattenNode()"
 	title="Flatten node and children-nodes, but keep an empty space"
 >
-	<img alt="flattenIcon" class="png-icon" src="../../../../assets/flatten.png" style="width: 17px; height: 17px" />
+	<img alt="flattenIcon" class="png-icon" src="../../../../../assets/flatten.png" style="width: 17px; height: 17px" />
 	FLATTEN
 </button>
 <button mat-button *ngIf="codeMapNode?.isFlattened" (click)="unFlattenNode()" title="Show node and children-nodes">
-	<img alt="flattenIcon" class="png-icon" src="../../../../assets/flatten.png" style="width: 17px; height: 17px" />
+	<img alt="flattenIcon" class="png-icon" src="../../../../../assets/flatten.png" style="width: 17px; height: 17px" />
 	SHOW
 </button>
