+++ visualization/app/codeCharta/state/effects/nodeContextMenu/flattenButtons/flattenButtons.module.ts
@@ -1,6 +1,6 @@
 import { CommonModule } from "@angular/common"
 import { NgModule } from "@angular/core"
-import { MaterialModule } from "../../../../material/material.module"
+import { MaterialModule } from "../../../../../material/material.module"
 import { FlattenButtonsComponent } from "./flattenButtons.component"
 
 @NgModule({
