+++ visualization/app/codeCharta/state/effects/nodeContextMenu/focusButtons/focusButtons.component.ts
@@ -1,11 +1,11 @@
 import { Component, Inject, Input } from "@angular/core"
 import { Observable, map } from "rxjs"
 
-import { CodeMapNode } from "../../../codeCharta.model"
-import { Store } from "../../../state/angular-redux/store"
-import { currentFocusedNodePathSelector } from "../../../state/store/dynamicSettings/focusedNodePath/currentFocused.selector"
-import { focusNode, unfocusAllNodes, unfocusNode } from "../../../state/store/dynamicSettings/focusedNodePath/focusedNodePath.actions"
-import { focusedNodePathSelector } from "../../../state/store/dynamicSettings/focusedNodePath/focusedNodePath.selector"
+import { CodeMapNode } from "../../../../codeCharta.model"
+import { Store } from "../../../angular-redux/store"
+import { currentFocusedNodePathSelector } from "../../../store/dynamicSettings/focusedNodePath/currentFocused.selector"
+import { focusNode, unfocusAllNodes, unfocusNode } from "../../../store/dynamicSettings/focusedNodePath/focusedNodePath.actions"
+import { focusedNodePathSelector } from "../../../store/dynamicSettings/focusedNodePath/focusedNodePath.selector"
 
 @Component({
 	selector: "cc-focus-buttons",
