+++ visualization/app/codeCharta/state/effects/nodeContextMenu/focusButtons/focusButtons.module.ts
@@ -1,7 +1,7 @@
 import { CommonModule } from "@angular/common"
 import { NgModule } from "@angular/core"
 
-import { MaterialModule } from "../../../../material/material.module"
+import { MaterialModule } from "../../../../../material/material.module"
 import { FocusButtonsComponent } from "./focusButtons.component"
 import { IsNodeFocusedPipe } from "./isNodeFocused.pipe"
