+++ visualization/app/codeCharta/state/effects/nodeContextMenu/highlightButtons/highlightButtons.component.spec.ts
@@ -2,11 +2,11 @@ import { render, screen } from "@testing-library/angular"
 import { TestBed } from "@angular/core/testing"
 import userEvent from "@testing-library/user-event"
 import { HighlightButtonsModule } from "./highlightButtons.module"
-import { ThreeSceneServiceToken } from "../../../services/ajs-upgraded-providers"
+import { ThreeSceneServiceToken } from "../../../../services/ajs-upgraded-providers"
 import { HighlightButtonsComponent } from "./highlightButtons.component"
-import { VALID_FILE_NODE_WITH_ID as codeMapNode } from "../../../util/dataMocks"
-import { State } from "../../../state/angular-redux/state"
-import { CodeMapNode } from "../../../codeCharta.model"
+import { VALID_FILE_NODE_WITH_ID as codeMapNode } from "../../../../util/dataMocks"
+import { State } from "../../../angular-redux/state"
+import { CodeMapNode } from "../../../../codeCharta.model"
 
 describe("flattenButtonsComponent", () => {
 	let threeSceneService
