+++ visualization/app/codeCharta/state/effects/nodeContextMenu/markFolderRow/markFolderRow.module.ts
@@ -1,6 +1,6 @@
 import { CommonModule } from "@angular/common"
 import { NgModule } from "@angular/core"
-import { ColorPickerModule } from "../../colorPicker/colorPicker.module"
+import { ColorPickerModule } from "../../../../ui/colorPicker/colorPicker.module"
 import { MarkFolderRowComponent } from "./markFolderRow.component"
 
 @NgModule({
