+++ visualization/app/codeCharta/state/effects/nodeContextMenu/nodeContextMenu.po.ts
@@ -1,4 +1,4 @@
-import { clickButtonOnPageElement } from "../../../puppeteer.helper"
+import { clickButtonOnPageElement } from "../../../../puppeteer.helper"
 
 export class NodeContextMenuPageObject {
 	async hasColorButtons() {
