+++ visualization/app/codeCharta/state/effects/nodeContextMenu/nodeContextMenu.service.ts
@@ -1,8 +1,8 @@
 import { Overlay, OverlayRef } from "@angular/cdk/overlay"
 import { ComponentPortal } from "@angular/cdk/portal"
 import { Inject, Injectable } from "@angular/core"
-import { Store } from "../../state/angular-redux/store"
-import { setRightClickedNodeData } from "../../state/store/appStatus/rightClickedNodeData/rightClickedNodeData.actions"
+import { Store } from "../../angular-redux/store"
+import { setRightClickedNodeData } from "../../store/appStatus/rightClickedNodeData/rightClickedNodeData.actions"
 import { NodeContextMenuCardComponent } from "./nodeContextMenuCard/nodeContextMenuCard.component"
 
 @Injectable({ providedIn: "root" })
