+++ visualization/app/codeCharta/state/effects/nodeContextMenu/nodeContextMenuCard/nodeContextMenuCard.component.ts
@@ -1,9 +1,9 @@
 import "./nodeContextMenuCard.component.scss"
 import { Component, Inject, OnInit } from "@angular/core"
-import { Store } from "../../../state/angular-redux/store"
+import { Store } from "../../../angular-redux/store"
 import { rightClickedCodeMapNodeSelector } from "../rightClickedCodeMapNode.selector"
 import { Observable } from "rxjs"
-import { CodeMapNode } from "../../../codeCharta.model"
+import { CodeMapNode } from "../../../../codeCharta.model"
 
 @Component({
 	selector: "cc-node-context-menu-card",
